Bugfix for remove_jobs_not_config
This fix is because the match for check if the queue is present in
reset_stats configs was checked inside the all dict (key) by queuename.

If you change the reset time for a determinated queue and this provokes
two different times to run the job for reset of stats for the same
queue.

The delete for a job should be done after every single check in
configuration for reset queues

This patch fix this case.