Bugfix to not run at the first time the reset stats when is add to queue:
Everytime is ran the job when is added for first time to queue.
This is because the scheduled_time should be set in UTC.
If are using a localtime in different time should produce a mismatch in
time.

Instead to use a schedule time with a repetead time with interval is run
the job in determined time using `enqueue_at` with time (in UTC)

Before run the `reset_stats_queue` checks this still present in configuration file
vs checking the queue.